Add Save Project Support !!!!!! -> added toString & fromSting on Parameters -> added position as FloatParameters in nodeBase -> NodeManagerUI modify position floatParameters and not ComponentBounds later NodeBaseUI reposition itself -> modified NodeManager clear() to ensure all removeListener are called -> added UNKNOWN enum in NodeType to implement typeto/FromString -> added default canvas with audioIn audioOut

+++ src/Blooper/Source/FloatRangeParameter.cpp
@@ -10,7 +10,7 @@
 
 #include "FloatRangeParameter.h"
 
-FloatRangeParameter::FloatRangeParameter(const String & niceName, const String &description, const int & initialValueMin, const int &initialValueMax, const int & minValue, const int & maxValue, bool enabled) :
+FloatRangeParameter::FloatRangeParameter(const String & niceName, const String &description, const float & initialValueMin, const float &initialValueMax, const float & minValue, const float & maxValue, bool enabled) :
 	Parameter(Type::RANGE,niceName, description, enabled), minValue(minValue), maxValue(maxValue)
 {
 	//base set values to min and max so they don't overlap when setting with setValues
